Shift through unsigned in bundled wavpack
Assembling a 32 bit value from bytes and shifting negative values are
undefined behaviour that UBSan reports as soon as anything decodes a
.wv file in a sanitizer build. Upstream WavPack fixed the same places
by changing the type the shift happens in: entropy_utils.c:374,
unpack.c:789, unpack.c:818 and common_utils.c:698.

Decoded samples of all 131 files in data/audio are bit identical
before and after.

+++ src/engine/external/wavpack/bits.c
@@ -85,7 +85,7 @@ void little_endian_to_native (void *data, char *format)
     while (*format) {
         switch (*format) {
             case 'L':
-                temp = cp [0] + ((int32_t) cp [1] << 8) + ((int32_t) cp [2] << 16) + ((int32_t) cp [3] << 24);
+                temp = (int32_t) ((uint32_t) cp [0] + ((uint32_t) cp [1] << 8) + ((uint32_t) cp [2] << 16) + ((uint32_t) cp [3] << 24));
                 * (int32_t *) cp = temp;
                 cp += 4;
                 break;
